move transaction hook to store
